CDAP-20657: Make Dataset / Dataframe first class citizens and reduce conversions to RDD
