+++ cdap-app-templates/cdap-etl/hydrator-spark-core-base/src/main/java/io/cdap/cdap/etl/spark/batch/BatchCollection.java
@@ -1,0 +1,36 @@
+/*
+ * Copyright © 2023 Cask Data, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not
+ * use this file except in compliance with the License. You may obtain a copy of
+ * the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+ * License for the specific language governing permissions and limitations under
+ * the License.
+ */
+
+package io.cdap.cdap.etl.spark.batch;
+
+import io.cdap.cdap.api.data.schema.Schema;
+import io.cdap.cdap.etl.spark.SparkCollection;
+
+/**
+ * An interface for Batch ETL collection. Such collections should be able to convert to
+ * {@link DataframeCollection} as soon as they contain
+ * {@link io.cdap.cdap.api.data.format.StructuredRecord}.
+ */
+public interface BatchCollection<T> extends SparkCollection<T> {
+
+  /**
+   * Returns a {@link DataframeCollection} for the the current colletion with given schema.
+   * It is assumed that current collection contains StructuredRecord rows.
+   * @param schema schema to convert StructuredRecod to {@link org.apache.spark.sql.Row}.
+   * @return a new DataframeCollection or this if it's already a DataframeCollection.
+   */
+  DataframeCollection toDataframeCollection(Schema schema);
+}
