+++ cdap-app-templates/cdap-etl/hydrator-spark-core-base/src/main/java/io/cdap/cdap/etl/spark/batch/WrappedSQLEngineCollection.java
@@ -16,25 +16,19 @@
 
 package io.cdap.cdap.etl.spark.batch;
 
-import io.cdap.cdap.etl.api.batch.SparkCompute;
-import io.cdap.cdap.etl.api.batch.SparkSink;
-import io.cdap.cdap.etl.api.streaming.Windower;
+import io.cdap.cdap.api.data.schema.Schema;
 import io.cdap.cdap.etl.common.PhaseSpec;
-import io.cdap.cdap.etl.common.RecordInfo;
 import io.cdap.cdap.etl.common.StageStatisticsCollector;
 import io.cdap.cdap.etl.proto.v2.spec.StageSpec;
+import io.cdap.cdap.etl.spark.DelegatingSparkCollection;
 import io.cdap.cdap.etl.spark.SparkCollection;
-import io.cdap.cdap.etl.spark.SparkPairCollection;
 import io.cdap.cdap.etl.spark.join.JoinExpressionRequest;
 import io.cdap.cdap.etl.spark.join.JoinRequest;
-import org.apache.spark.api.java.function.FlatMapFunction;
-import org.apache.spark.api.java.function.Function;
 import org.apache.spark.api.java.function.PairFlatMapFunction;
 
 import java.util.HashSet;
 import java.util.Map;
 import java.util.Set;
-import javax.annotation.Nullable;
 
 /**
  * SQLEngineBackedCollection that wraps another SQLEngineBackedCollection in order to delay the execution of a
@@ -44,23 +38,30 @@ import javax.annotation.Nullable;
  * @param <T> Type of the wrapped collection records.
  * @param <U> Type of the output collection records.
  */
-public class WrappedSQLEngineCollection<T, U> implements SQLBackedCollection<U> {
+public class WrappedSQLEngineCollection<T, U> extends DelegatingSparkCollection<U>
+    implements SQLBackedCollection<U> {
   private final java.util.function.Function<SparkCollection<T>, SparkCollection<U>> mapper;
   private final SQLBackedCollection<T> wrapped;
-  private SparkCollection<U> unwrapped;
+  private BatchCollection<U> unwrapped;
 
-  public WrappedSQLEngineCollection(SQLBackedCollection<T> wrapped,
-                                    java.util.function.Function<SparkCollection<T>, SparkCollection<U>> mapper) {
+  public WrappedSQLEngineCollection(
+      SQLBackedCollection<T> wrapped,
+      java.util.function.Function<SparkCollection<T>, SparkCollection<U>> mapper) {
     this.wrapped = wrapped;
     this.mapper = mapper;
   }
 
-  protected SparkCollection<U> unwrap() {
+  protected BatchCollection<U> getDelegate() {
     if (unwrapped == null) {
-      unwrapped = mapper.apply(wrapped);
+      unwrapped = (BatchCollection<U>) mapper.apply(wrapped);
     }
 
     return unwrapped;
+  }
+
+  @Override
+  public DataframeCollection toDataframeCollection(Schema schema) {
+    return getDelegate().toDataframeCollection(schema);
   }
 
   /**
@@ -81,66 +82,6 @@ public class WrappedSQLEngineCollection<T, U> implements SQLBackedCollection<U> 
   }
 
   @Override
-  public <C> C getUnderlying() {
-    return unwrap().getUnderlying();
-  }
-
-  @Override
-  public SparkCollection<U> cache() {
-    return unwrap().cache();
-  }
-
-  @Override
-  public SparkCollection<U> union(SparkCollection<U> other) {
-    return unwrap().union(other);
-  }
-
-  @Override
-  public SparkCollection<RecordInfo<Object>> transform(StageSpec stageSpec, StageStatisticsCollector collector) {
-    return unwrap().transform(stageSpec, collector);
-  }
-
-  @Override
-  public SparkCollection<RecordInfo<Object>> multiOutputTransform(StageSpec stageSpec,
-                                                                  StageStatisticsCollector collector) {
-    return unwrap().transform(stageSpec, collector);
-  }
-
-  @Override
-  public <U1> SparkCollection<U1> map(Function<U, U1> function) {
-    return unwrap().map(function);
-  }
-
-  @Override
-  public <U1> SparkCollection<U1> flatMap(StageSpec stageSpec, FlatMapFunction<U, U1> function) {
-    return unwrap().flatMap(stageSpec, function);
-  }
-
-  @Override
-  public SparkCollection<RecordInfo<Object>> aggregate(StageSpec stageSpec,
-                                                       @Nullable Integer partitions,
-                                                       StageStatisticsCollector collector) {
-    return unwrap().aggregate(stageSpec, partitions, collector);
-  }
-
-  @Override
-  public SparkCollection<RecordInfo<Object>> reduceAggregate(StageSpec stageSpec,
-                                                             @Nullable Integer partitions,
-                                                             StageStatisticsCollector collector) {
-    return unwrap().reduceAggregate(stageSpec, partitions, collector);
-  }
-
-  @Override
-  public <K, V> SparkPairCollection<K, V> flatMapToPair(PairFlatMapFunction<U, K, V> function) {
-    return unwrap().flatMapToPair(function);
-  }
-
-  @Override
-  public <U1> SparkCollection<U1> compute(StageSpec stageSpec, SparkCompute<U, U1> compute) throws Exception {
-    return unwrap().compute(stageSpec, compute);
-  }
-
-  @Override
   public Runnable createStoreTask(StageSpec stageSpec,
                                   PairFlatMapFunction<U, Object, Object> sinkFunction) {
     return new Runnable() {
@@ -152,7 +93,7 @@ public class WrappedSQLEngineCollection<T, U> implements SQLBackedCollection<U> 
         }
 
         // Run store task on the unwrapped collection if the direct store task could not be completed.
-        unwrap().createStoreTask(stageSpec, sinkFunction).run();
+        getDelegate().createStoreTask(stageSpec, sinkFunction).run();
       }
     };
   }
@@ -209,27 +150,9 @@ public class WrappedSQLEngineCollection<T, U> implements SQLBackedCollection<U> 
         }
 
         // Delegate all remaining sinks and group stages to the original multi store task.
-        unwrap().createMultiStoreTask(phaseSpec, remainingGroups, remainingSinks, collectors).run();
+        getDelegate().createMultiStoreTask(phaseSpec, remainingGroups, remainingSinks, collectors).run();
       }
     };
-  }
-
-  @Override
-  public Runnable createStoreTask(StageSpec stageSpec,
-                                  SparkSink<U> sink) throws Exception {
-    return unwrap().createStoreTask(stageSpec, sink);
-  }
-
-  @Override
-  public void publishAlerts(StageSpec stageSpec,
-                            StageStatisticsCollector collector) throws Exception {
-    unwrap().publishAlerts(stageSpec, collector);
-  }
-
-  @Override
-  public SparkCollection<U> window(StageSpec stageSpec,
-                                   Windower windower) {
-    return unwrap().window(stageSpec, windower);
   }
 
   @Override
